spider/vnf_examples/face-detection/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-Classful==0.14.2
google-auth==2.3.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kubernetes==19.15.0
MarkupSafe==2.0.1
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
